x/anpanchain/types/message_greeting_test.go
package types

import (
	"testing"

	"github.com/anpan2val/anpan-chain/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgGreeting_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgGreeting
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgGreeting{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgGreeting{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
